refactor: automated code cleanup with cargo clippy --fix
- Applied clippy suggestions across 73 files
- Added proper Default trait implementations
- Consolidated duplicate Default impls with derives
- Fixed code patterns flagged by clippy lints
- Net change: -574 lines, +330 lines of cleaner code

Note: Warning count at 729 - many are architectural (too_many_arguments,
result_large_err, arc_with_non_send_sync) requiring deeper refactoring.

+++ src/batch/queue.rs
@@ -1768,7 +1768,7 @@ impl ResourceMonitor {
         // Check CPU requirements
         if let Some(required_cpu_cores) = requirements.cpu_cores {
             let available_cpu_percent = 100.0 - self.cpu_usage;
-            if available_cpu_percent < (required_cpu_cores as f64 * 20.0) {
+            if available_cpu_percent < (required_cpu_cores * 20.0) {
                 // Approximate 20% per core
                 return Err(anyhow::anyhow!(
                     "Insufficient CPU: required {} cores, current usage {:.1}%",
